refactor cloudwatch code
